@@ -1,21 +1,22 @@
 {
   "license": "MIT",
+  "version": "0.4.7",
   "scripts": {
     "clean": "rm -rf build",
     "build-examples": "npm run clean && webpack",
     "deploy-demos": "git checkout gh-pages && git merge master --no-edit && npm run build-examples && git add -A && git commit -m 'Update demos' && git push && git checkout master",
     "build": "npm run clean && babel --out-dir build lib",
     "dev": "npm start",
-    "test": "testem",
+    "test": "jest",
+    "test-dev": "jest --watch",
     "start": "webpack-dev-server --hot",
-    "prepublish": "npm run build",
-    "postversion": "git push && git push --tags && npm publish"
+    "postversion": "npm run build && git push && git push --tags && npm publish && release"
   },
   "dependencies": {
+    "babel-jest": "^19.0.0",
     "css-vendor": "^0.3.1",
     "debug": "^2.1.3",
-    "lodash.throttle": "^3.0.3",
-    "object.assign": "^4.0.1"
+    "lodash.throttle": "^3.0.3"
   },
   "repository": {
     "type": "git",
@@ -24,32 +25,35 @@
   "author": "Jason Kuhrt <jasonkuhrt@me.com> (http://jasonkuhrt.com/)",
   "name": "react-popover",
   "homepage": "https://github.com/littlebits/react-popover",
-  "version": "0.4.0",
   "main": "build/index.js",
   "description": "A specification backed popover for react",
   "devDependencies": {
     "babel-cli": "^6.3.15",
     "babel-core": "^6.3.15",
+    "babel-eslint": "^7.1.1",
     "babel-loader": "^6.2.0",
     "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-stage-1": "^6.3.13",
     "bluebird": "^3.0.5",
     "bootstrap-webpack-plugin": "^0.3.0",
     "chai": "^3.2.0",
     "classnames": "^2.2.1",
-    "coffee-loader": "^0.7.2",
-    "coffee-script": "^1.9.1",
     "css-loader": "^0.17.0",
     "cssnext-loader": "^1.0.1",
-    "eslint-config-littlebits": "^0.2.1",
-    "eslint-plugin-react": "^3.11.3",
+    "eslint": "^3.17.1",
+    "eslint-config-littlebits": "^0.6.2",
+    "jest": "^19.0.2",
+    "jest-cli": "^19.0.2",
     "ramda": "^0.18.0",
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
     "react-draggable": "^1.1.1",
     "react-hot-loader": "^1.2.4",
     "react-tappable": "^0.7.1",
+    "react-test-renderer": "^15.4.2",
+    "release": "^1.1.7",
     "style-loader": "^0.12.3",
-    "testem": "^0.9.6",
     "webpack": "^1.12.1",
     "webpack-dev-server": "^1.10.1"
   },
@@ -57,20 +61,19 @@
     "url": "https://github.com/littlebits/react-popover/issues"
   },
   "eslintConfig": {
+    "parser": "babel-eslint",
     "extends": [
       "littlebits"
     ],
-    "plugins": [
-      "react"
-    ],
-    "ecmaFeatures": {
-      "jsx": true
+    "env": {
+      "jest": true
     }
   },
   "babel": {
     "presets": [
       "es2015",
-      "react"
+      "react",
+      "stage-1"
     ]
   }
 }
