@@ -1,17 +1,18 @@
-import { createElement, createClass, DOM as E, PropTypes as T } from 'react'
-import { findDOMNode } from 'react-dom'
-import Debug from 'debug'
-import * as resizeEvent from './on-resize'
-import * as Layout from './layout'
-import throttle from 'lodash.throttle'
-const cssVendor = require('css-vendor')
-import { isServer, window } from './platform'
-import ReactLayerMixin from './react-layer-mixin'
-import { arrayify, assign, clientOnly } from './utils'
-import PopoverTip from './popover-tip'
+import { createElement, createClass, DOM as E, PropTypes as T } from "react"
+import { findDOMNode } from "react-dom"
+import Debug from "debug"
+import throttle from "lodash.throttle"
+import * as cssVendor from "css-vendor"
+import resizeEvent from "./on-resize"
+import Layout from "./layout"
+import ReactLayerMixin from "./react-layer-mixin"
+import { isServer, window, document } from "./platform"
+import { arrayify, clientOnly } from "./utils"
+import Tip from "./tip"
 
 
-const log = Debug(`react-popover`)
+
+const log = Debug("react-popover")
 
 const supportedCSSValue = clientOnly(cssVendor.supportedValue)
 
@@ -28,55 +29,50 @@ const cssvalue = (prop, value) => (
 )
 
 const coreStyle = {
-  position: `absolute`,
+  position: "absolute",
   top: 0,
   left: 0,
-  display: cssvalue(`display`, `flex`),
+  display: cssvalue("display", "flex"),
 }
 
 const faces = {
-  above: `down`,
-  right: `left`,
-  below: `up`,
-  left: `right`,
+  above: "down",
+  right: "left",
+  below: "up",
+  left: "right",
 }
 
 /* Flow mappings. Each map maps the flow domain to another domain. */
 
 const flowToTipTranslations = {
-  row: `translateY`,
-  column: `translateX`,
+  row: "translateY",
+  column: "translateX",
 }
 
 const flowToPopoverTranslations = {
-  row: `translateX`,
-  column: `translateY`,
+  row: "translateX",
+  column: "translateY",
 }
 
 
 
 const Popover = createClass({
-  displayName: `popover`,
-  mixins: [ReactLayerMixin()],
+  displayName: "popover",
   propTypes: {
     body: T.node.isRequired,
     children: T.element.isRequired,
-    preferPlace: T.oneOf(Layout.validTypeValues),
-    place: T.oneOf(Layout.validTypeValues),
-    tipSize: T.number,
-    refreshIntervalMs: T.oneOfType([ T.number, T.bool ]),
-    isOpen: T.bool,
-    onOuterAction: T.func,
+    className: T.string,
     enterExitTransitionDurationMs: T.number,
+    isOpen: T.bool,
+    offset: T.number,
+    place: T.oneOf(Layout.validTypeValues),
+    preferPlace: T.oneOf(Layout.validTypeValues),
+    refreshIntervalMs: T.oneOfType([ T.number, T.bool ]),
+    style: T.object,
+    tipSize: T.number,
+    onOuterAction: T.func,
   },
-  getInitialState () {
-    return {
-      standing: `above`,
-      exited: !this.props.isOpen, // for animation-dependent rendering, should popover close/open?
-      exiting: false, // for tracking in-progress animations
-      toggle: false, // for business logic tracking, should popover close/open?
-    }
-  },
+  mixins: [ReactLayerMixin()],
   getDefaultProps () {
     return {
       tipSize: 7,
@@ -90,8 +86,40 @@ const Popover = createClass({
       refreshIntervalMs: 200,
     }
   },
-  checkTargetReposition () {
-    if (this.measureTargetBounds()) this.resolvePopoverLayout()
+  getInitialState () {
+    return {
+      standing: "above",
+      exited: !this.props.isOpen, // for animation-dependent rendering, should popover close/open?
+      exiting: false, // for tracking in-progress animations
+      toggle: false, // for business logic tracking, should popover close/open?
+    }
+  },
+  componentDidMount () {
+    this.targetEl = findDOMNode(this)
+    if (this.props.isOpen) this.enter()
+  },
+  componentWillReceiveProps (propsNext) {
+    //log(`Component received props!`, propsNext)
+    const willOpen = !this.props.isOpen && propsNext.isOpen
+    const willClose = this.props.isOpen && !propsNext.isOpen
+
+    if (willOpen) this.open()
+    else if (willClose) this.close()
+
+  },
+  componentDidUpdate (propsPrev, statePrev) {
+    //log(`Component did update!`)
+    const didOpen = !statePrev.toggle && this.state.toggle
+    const didClose = statePrev.toggle && !this.state.toggle
+
+    if (didOpen) this.enter()
+    else if (didClose) this.exit()
+  },
+  componentWillUnmount () {
+    /* If the Popover was never opened then then tracking
+    initialization never took place and so calling untrack
+    would be an error. Also see issue 55. */
+    if (this.hasTracked) this.untrackPopover()
   },
   resolvePopoverLayout () {
 
@@ -116,20 +144,20 @@ const Popover = createClass({
     way at any time for any arbitrary trigger. There may be value in investigating the
     http://overconstrained.io community for its general layout system via the
     constraint-solver Cassowary. */
-    if (this.zone) this.size[this.zone.flow === `row` ? `h` : `w`] += this.props.tipSize
+    if (this.zone) this.size[this.zone.flow === "row" ? "h" : "w"] += this.props.tipSize
     const zone = Layout.pickZone(pickerSettings, this.frameBounds, this.targetBounds, this.size)
-    if (this.zone) this.size[this.zone.flow === `row` ? `h` : `w`] -= this.props.tipSize
+    if (this.zone) this.size[this.zone.flow === "row" ? "h" : "w"] -= this.props.tipSize
 
     const tb = this.targetBounds
     this.zone = zone
-    log(`zone`, zone)
+    log("zone", zone)
 
     this.setState({
       standing: zone.standing,
     })
 
     const axis = Layout.axes[zone.flow]
-    log(`axes`, axis)
+    log("axes", axis)
 
     const dockingEdgeBufferLength = Math.round(getComputedStyle(this.bodyEl).borderRadius.slice(0, -2)) || 0
     const scrollSize = Layout.El.calcScrollSize(this.frameEl)
@@ -164,29 +192,33 @@ const Popover = createClass({
     /* If the popover dose not fit into frameCrossLength then just position it to the `frameCrossStart`.
     popoverCrossLength` will now be forced to overflow into the `Frame` */
     if (pos.crossLength > frameCrossLength) {
-      log(`popoverCrossLength does not fit frame.`)
+      log("popoverCrossLength does not fit frame.")
       pos[axis.cross.start] = 0
 
     /* If the `popoverCrossStart` is forced beyond some threshold of `targetCrossLength` then bound
     it (`popoverCrossStart`). */
-
     } else if (tb[axis.cross.end] < hangingBufferLength) {
-      log(`popoverCrossStart cannot hang any further without losing target.`)
+      log("popoverCrossStart cannot hang any further without losing target.")
       pos[axis.cross.start] = tb[axis.cross.end] - hangingBufferLength
+
+    /* checking if the cross start of the target area is within the frame and it makes sense
+    to try fitting popover into the frame. */
+    } else if (tb[axis.cross.start] > frameCrossInnerEnd) {
+      log("popoverCrossStart cannot hang any further without losing target.")
+      pos[axis.cross.start] = tb[axis.cross.start] - this.size[axis.cross.size]
 
     /* If the `popoverCrossStart` does not fit within the inner frame (honouring buffers) then
     just center the popover in the remaining `frameCrossLength`. */
-
     } else if (pos.crossLength > frameCrossInnerLength) {
-      log(`popoverCrossLength does not fit within buffered frame.`)
+      log("popoverCrossLength does not fit within buffered frame.")
       pos[axis.cross.start] = (frameCrossLength - pos.crossLength) / 2
 
     } else if (popoverCrossStart < frameCrossInnerStart) {
-      log(`popoverCrossStart cannot reverse without exceeding frame.`)
+      log("popoverCrossStart cannot reverse without exceeding frame.")
       pos[axis.cross.start] = frameCrossInnerStart
 
     } else if (popoverCrossEnd > frameCrossInnerEnd) {
-      log(`popoverCrossEnd cannot travel without exceeding frame.`)
+      log("popoverCrossEnd cannot travel without exceeding frame.")
       pos[axis.cross.start] = pos[axis.cross.start] - (pos[axis.cross.end] - frameCrossInnerEnd)
     }
 
@@ -202,13 +234,13 @@ const Popover = createClass({
     recalculate layout. */
 
     this.containerEl.style.flexFlow = zone.flow
-    this.containerEl.style[jsprefix(`FlexFlow`)] = this.containerEl.style.flexFlow
+    this.containerEl.style[jsprefix("FlexFlow")] = this.containerEl.style.flexFlow
     this.bodyEl.style.order = zone.order
-    this.bodyEl.style[jsprefix(`Order`)] = this.bodyEl.style.order
+    this.bodyEl.style[jsprefix("Order")] = this.bodyEl.style.order
 
     /* Apply Absolute Positioning. */
 
-    log(`pos`, pos)
+    log("pos", pos)
     this.containerEl.style.top = `${pos.y}px`
     this.containerEl.style.left = `${pos.x}px`
 
@@ -218,7 +250,7 @@ const Popover = createClass({
       /* Get the absolute tipCrossCenter. Tip is positioned relative to containerEl
       but it aims at targetCenter which is positioned relative to frameEl... we
       need to cancel the containerEl positioning so as to hit our intended position. */
-      Layout.centerOfBoundsFromBounds(zone.flow, `cross`, tb, pos)
+      Layout.centerOfBoundsFromBounds(zone.flow, "cross", tb, pos)
 
       /* centerOfBounds does not account for scroll so we need to manually add that
       here. */
@@ -235,7 +267,10 @@ const Popover = createClass({
     }
 
     this.tipEl.style.transform = `${flowToTipTranslations[zone.flow]}(${tipCrossPos}px)`
-    this.tipEl.style[jsprefix(`Transform`)] = this.tipEl.style.transform
+    this.tipEl.style[jsprefix("Transform")] = this.tipEl.style.transform
+  },
+  checkTargetReposition () {
+    if (this.measureTargetBounds()) this.resolvePopoverLayout()
   },
   measurePopoverSize () {
     this.size = Layout.El.calcSize(this.containerEl)
@@ -250,47 +285,21 @@ const Popover = createClass({
     this.targetBounds = newTargetBounds
     return true
   },
-  componentDidMount () {
-    this.targetEl = findDOMNode(this)
-    if (this.props.isOpen) this.enter()
-  },
-  componentWillReceiveProps (propsNext) {
-    //log(`Component received props!`, propsNext)
-    const willOpen = !this.props.isOpen && propsNext.isOpen
-    const willClose = this.props.isOpen && !propsNext.isOpen
-
-    if (willOpen) this.open()
-    else if (willClose) this.close()
-
-  },
   open () {
     if (this.state.exiting) this.animateExitStop()
     this.setState({ toggle: true, exited: false })
   },
   close () {
-    /* TODO?: we currently do not setup any `entering` state flag because
-    nothing would really need to depend on it. Stopping animations is currently nothing
-    more than clearing some timeouts which are safe to clear even if undefined. The
-    primary reason for `exiting` state is for the `layerRender` logic. */
-    this.animateEnterStop()
     this.setState({ toggle: false })
-  },
-  componentDidUpdate (propsPrev, statePrev) {
-    //log(`Component did update!`)
-    const didOpen = !statePrev.toggle && this.state.toggle
-    const didClose = statePrev.toggle && !this.state.toggle
-
-    if (didOpen) this.enter()
-    else if (didClose) this.exit()
   },
   enter () {
     if (isServer) return
-    log(`enter!`)
+    log("enter!")
     this.trackPopover()
     this.animateEnter()
   },
   exit () {
-    log(`exit!`)
+    log("exit!")
     this.animateExit()
     this.untrackPopover()
   },
@@ -304,7 +313,7 @@ const Popover = createClass({
     this.exitingAnimationTimer2 = setTimeout(() => {
       setTimeout(() => {
         this.containerEl.style.transform = `${flowToPopoverTranslations[this.zone.flow]}(${this.zone.order * 50}px)`
-        this.containerEl.style.opacity = `0`
+        this.containerEl.style.opacity = "0"
       }, 0)
     }, 0)
 
@@ -312,31 +321,25 @@ const Popover = createClass({
       this.setState({ exited: true, exiting: false })
     }, this.props.enterExitTransitionDurationMs)
   },
-  animateEnterStop () {
-    clearTimeout(this.enteringAnimationTimer1)
-    clearTimeout(this.enteringAnimationTimer2)
-  },
   animateEnter () {
     /* Prepare `entering` style so that we can then animate it toward `entered`. */
 
     this.containerEl.style.transform = `${flowToPopoverTranslations[this.zone.flow]}(${this.zone.order * 50}px)`
-    this.containerEl.style[jsprefix(`Transform`)] = this.containerEl.style.transform
-    this.containerEl.style.opacity = `0`
+    this.containerEl.style[jsprefix("Transform")] = this.containerEl.style.transform
+    this.containerEl.style.opacity = "0"
 
     /* After initial layout apply transition animations. */
+    /* Hack: http://stackoverflow.com/questions/3485365/how-can-i-force-webkit-to-redraw-repaint-to-propagate-style-changes */
+    this.containerEl.offsetHeight
 
-    this.enteringAnimationTimer1 = setTimeout(() => {
-      this.tipEl.style.transition = `transform 150ms ease-in`
-      this.tipEl.style[jsprefix(`Transition`)] = `${cssprefix(`transform`)} 150ms ease-in`
-      this.containerEl.style.transitionProperty = `top, left, opacity, transform`
-      this.containerEl.style.transitionDuration = `500ms`
-      this.containerEl.style.transitionTimingFunction = `cubic-bezier(0.230, 1.000, 0.320, 1.000)`
-      this.enteringAnimationTimer2 = setTimeout(() => {
-        this.containerEl.style.opacity = `1`
-        this.containerEl.style.transform = `translateY(0)`
-        this.containerEl.style[jsprefix(`Transform`)] = this.containerEl.style.transform
-      }, 0)
-    }, 0)
+    this.tipEl.style.transition = "transform 150ms ease-in"
+    this.tipEl.style[jsprefix("Transition")] = `${cssprefix("transform")} 150ms ease-in`
+    this.containerEl.style.transitionProperty = "top, left, opacity, transform"
+    this.containerEl.style.transitionDuration = "500ms"
+    this.containerEl.style.transitionTimingFunction = "cubic-bezier(0.230, 1.000, 0.320, 1.000)"
+    this.containerEl.style.opacity = "1"
+    this.containerEl.style.transform = "translateY(0)"
+    this.containerEl.style[jsprefix("Transform")] = this.containerEl.style.transform
   },
   trackPopover () {
     const minScrollRefreshIntervalMs = 200
@@ -345,14 +348,15 @@ const Popover = createClass({
     /* Get references to DOM elements. */
 
     this.containerEl = findDOMNode(this.layerReactComponent)
-    this.bodyEl = this.containerEl.querySelector(`.Popover-body`)
-    this.tipEl = this.containerEl.querySelector(`.Popover-tip`)
+    this.bodyEl = this.containerEl.querySelector(".Popover-body")
+    this.tipEl = this.containerEl.querySelector(".Popover-tip")
 
     /* Note: frame is hardcoded to window now but we think it will
     be a nice feature in the future to allow other frames to be used
     such as local elements that further constrain the popover`s world. */
 
     this.frameEl = window
+    this.hasTracked = true
 
     /* Set a general interval for checking if target position changed. There is no way
     to know this information without polling. */
@@ -371,7 +375,7 @@ const Popover = createClass({
     this.onPopoverResize = throttle(this.onPopoverResize, minResizeRefreshIntervalMs)
     this.onTargetResize = throttle(this.onTargetResize, minResizeRefreshIntervalMs)
 
-    this.frameEl.addEventListener(`scroll`, this.onFrameScroll)
+    this.frameEl.addEventListener("scroll", this.onFrameScroll)
     resizeEvent.on(this.frameEl, this.onFrameResize)
     resizeEvent.on(this.containerEl, this.onPopoverResize)
     resizeEvent.on(this.targetEl, this.onTargetResize)
@@ -379,8 +383,8 @@ const Popover = createClass({
     /* Track user actions on the page. Anything that occurs _outside_ the Popover boundaries
     should close the Popover. */
 
-    window.addEventListener(`mousedown`, this.checkForOuterAction)
-    window.addEventListener(`touchstart`, this.checkForOuterAction)
+    document.addEventListener("mousedown", this.checkForOuterAction)
+    document.addEventListener("touchstart", this.checkForOuterAction)
 
     /* Kickstart layout at first boot. */
 
@@ -394,51 +398,48 @@ const Popover = createClass({
       !this.containerEl.contains(event.target) &&
       !this.targetEl.contains(event.target)
     )
-    if (isOuterAction) this.props.onOuterAction()
+    if (isOuterAction) this.props.onOuterAction(event)
   },
   untrackPopover () {
     clearInterval(this.checkLayoutInterval)
-    this.frameEl.removeEventListener(`scroll`, this.onFrameScroll)
+    this.frameEl.removeEventListener("scroll", this.onFrameScroll)
     resizeEvent.off(this.frameEl, this.onFrameResize)
     resizeEvent.off(this.containerEl, this.onPopoverResize)
     resizeEvent.off(this.targetEl, this.onTargetResize)
-    window.removeEventListener(`mousedown`, this.checkForOuterAction)
-    window.removeEventListener(`touchstart`, this.checkForOuterAction)
+    document.removeEventListener("mousedown", this.checkForOuterAction)
+    document.removeEventListener("touchstart", this.checkForOuterAction)
   },
   onTargetResize () {
-    log(`Recalculating layout because _target_ resized!`)
+    log("Recalculating layout because _target_ resized!")
     this.measureTargetBounds()
     this.resolvePopoverLayout()
   },
   onPopoverResize () {
-    log(`Recalculating layout because _popover_ resized!`)
+    log("Recalculating layout because _popover_ resized!")
     this.measurePopoverSize()
     this.resolvePopoverLayout()
   },
   onFrameScroll () {
-    log(`Recalculating layout because _frame_ scrolled!`)
+    log("Recalculating layout because _frame_ scrolled!")
     this.measureTargetBounds()
     this.resolvePopoverLayout()
   },
   onFrameResize () {
-    log(`Recalculating layout because _frame_ resized!`)
+    log("Recalculating layout because _frame_ resized!")
     this.measureFrameBounds()
     this.resolvePopoverLayout()
   },
   measureFrameBounds () {
     this.frameBounds = Layout.El.calcBounds(this.frameEl)
   },
-  componentWillUnmount () {
-    clearInterval(this.checkLayoutInterval)
-  },
   renderLayer () {
     if (this.state.exited) return null
 
-    const { className = ``, style = {}} = this.props
+    const { className = "", style = {}} = this.props
 
     const popoverProps = {
       className: `Popover ${className}`,
-      style: assign({}, coreStyle, style),
+      style: { ...coreStyle, ...style }
     }
 
     const tipProps = {
@@ -455,8 +456,8 @@ const Popover = createClass({
 
     return (
       E.div(popoverProps,
-        E.div({ className: `Popover-body` }, ...popoverBody),
-        createElement(PopoverTip, tipProps)
+        E.div({ className: "Popover-body" }, ...popoverBody),
+        createElement(Tip, tipProps)
       )
     )
   },
@@ -467,6 +468,6 @@ const Popover = createClass({
 
 
 
-export {
-  Popover as default,
-}
+// Support for CJS
+// http://stackoverflow.com/questions/33505992/babel-6-changes-how-it-exports-default
+module.exports = Popover
