@@ -1,5 +1,5 @@
-import { window } from './platform'
-import { find, equalRecords } from './utils'
+import { window } from "./platform"
+import { find, equalRecords } from "./utils"
 
 
 
@@ -15,14 +15,14 @@ const axes = {
 }
 
 axes.row.main = {
-  start: `x`,
-  end: `x2`,
-  size: `w`,
+  start: "x",
+  end: "x2",
+  size: "w",
 }
 axes.row.cross = {
-  start: `y`,
-  end: `y2`,
-  size: `h`,
+  start: "y",
+  end: "y2",
+  size: "h",
 }
 axes.column.main = axes.row.cross
 axes.column.cross = axes.row.main
@@ -30,9 +30,9 @@ axes.column.cross = axes.row.main
 
 
 const types = [
-  { name: `side`, values: [ `start`, `end` ]},
-  { name: `standing`, values: [ `above`, `right`, `below`, `left` ]},
-  { name: `flow`, values: [ `column`, `row` ]},
+  { name: "side", values: [ "start", "end" ]},
+  { name: "standing", values: [ "above", "right", "below", "left" ]},
+  { name: "flow", values: [ "column", "row" ]},
 ]
 
 const validTypeValues = (
@@ -57,11 +57,11 @@ const centerOfBoundsFromBounds = (flow, axis, boundsTo, boundsFrom) => (
 const place = (flow, axis, align, bounds, size) => {
   const axisProps = axes[flow][axis]
   return (
-    align === `center`
+    align === "center"
       ? centerOfBounds(flow, axis, bounds) - centerOfSize(flow, axis, size)
-    : align === `end`
+    : align === "end"
       ? bounds[axisProps.end]
-    : align === `start`
+    : align === "start"
       /* DOM rendering unfolds leftward. Therefore if the slave is positioned before
       the master then the slave`s position must in addition be pulled back
       by its [the slave`s] own length. */
@@ -109,10 +109,10 @@ El.calcSize = (el) => (
 
 El.calcScrollSize = (el) => (
   el === window ?
-    {
-      w: el.scrollX || el.pageXOffset,
-      h: el.scrollY || el.pageYOffset,
-    } :
+  {
+    w: el.scrollX || el.pageXOffset,
+    h: el.scrollY || el.pageYOffset,
+  } :
     { w: el.scrollLeft, h: el.scrollTop }
 )
 
@@ -135,11 +135,11 @@ const getPreferenceType = (preference) => (
 /* Dimension Fit Checks */
 
 const fitWithinChecker = (dimension) => (domainSize, itemSize) => (
-  domainSize[dimension] > itemSize[dimension]
+  domainSize[dimension] >= itemSize[dimension]
 )
 
-const doesWidthFitWithin = fitWithinChecker(`w`)
-const doesHeightFitWithin = fitWithinChecker(`h`)
+const doesWidthFitWithin = fitWithinChecker("w")
+const doesHeightFitWithin = fitWithinChecker("h")
 
 const doesFitWithin = (domainSize, itemSize) => (
   doesWidthFitWithin(domainSize, itemSize)
@@ -152,7 +152,7 @@ const doesFitWithin = (domainSize, itemSize) => (
 
 const createPreferenceError = (givenValue) => (
   new Error(
-    `The given layout placement of "${givenValue}" is not a valid choice. Valid choices are: ${validTypeValues.join(` | `)}.`
+    `The given layout placement of "${givenValue}" is not a valid choice. Valid choices are: ${validTypeValues.join(" | ")}.`
   )
 )
 
@@ -160,20 +160,30 @@ const createPreferenceError = (givenValue) => (
 
 
 
-/* Algorithm for picking the best fitting zone for popover. The current technique will loop through all zones picking the last one that fits. If
-none fit the last one is selected.
-
-TODO In the case that none fit we should pick the least-not-fitting zone. */
+/* Algorithm for picking the best fitting zone for popover. The current technique will loop through all zones picking the last one that fits.
+In the case that none fit we should pick the least-not-fitting zone. */
 
 const pickZone = (opts, frameBounds, targetBounds, size) => {
   const t = targetBounds
   const f = frameBounds
   const zones = [
-    { side: `start`, standing: `above`, flow: `column`, order: -1, w: f.x2, h: t.y },
-    { side: `end`, standing: `right`, flow: `row`, order: 1, w: (f.x2 - t.x2), h: f.y2 },
-    { side: `end`, standing: `below`, flow: `column`, order: 1, w: f.x2, h: (f.y2 - t.y2) },
-    { side: `start`, standing: `left`, flow: `row`, order: -1, w: t.x, h: f.y2 },
+    { side: "start", standing: "above", flow: "column", order: -1, w: f.x2, h: t.y },
+    { side: "end", standing: "right", flow: "row", order: 1, w: (f.x2 - t.x2), h: f.y2 },
+    { side: "end", standing: "below", flow: "column", order: 1, w: f.x2, h: (f.y2 - t.y2) },
+    { side: "start", standing: "left", flow: "row", order: -1, w: t.x, h: f.y2 },
   ]
+
+  /* Order the zones by the amount of popup that would be cut out if that zone is used.
+     The first one in the array is the one that cuts the least amount.
+
+     const area = size.w * size.h  // Popup area is constant and it does not change the order
+  */
+  zones.forEach((z) => {
+    // TODO Update to satisfy linter
+    // eslint-disable-next-line no-param-reassign
+    z.cutOff = /* area */ - Math.max(0, Math.min(z.w,size.w)) * Math.max(0, Math.min(z.h,size.h))
+  })
+  zones.sort((a,b) => a.cutOff - b.cutOff)
 
   const availZones = zones.filter((zone) => (
     doesFitWithin(zone, size)
@@ -190,17 +200,27 @@ const pickZone = (opts, frameBounds, targetBounds, size) => {
 
   /* If the preferred side is part of the available zones, use that otherwise
   pick the largest available zone. If there are no available zones, pick the
-  largest zone. TODO: logic that executes picking based on largest option. */
+  largest zone. */
 
   if (opts.preferPlace) {
     const preferenceType = getPreferenceType(opts.preferPlace)
     if (!preferenceType) throw createPreferenceError(opts.preferPlace)
+
+    // Try to fit first in zone where the pop up fit completely
     const preferredAvailZones = availZones.filter((zone) => (
       zone[preferenceType] === opts.preferPlace
     ))
     if (preferredAvailZones.length) return preferredAvailZones[0]
+
+    // If there are not areas where the pop up fit completely, it uses the preferred ones
+    // in order from the one the fit better
+    const preferredZones = zones.filter((zone) => (
+      zone[preferenceType] === opts.preferPlace
+    ))
+    if (!availZones.length && preferredZones.length) return preferredZones[0]
   }
 
+  // Return a zone that fit completely or the one that fit the best
   return availZones.length ? availZones[0] : zones[0]
 }
 
@@ -211,10 +231,10 @@ const pickZone = (opts, frameBounds, targetBounds, size) => {
 const calcRelPos = (zone, masterBounds, slaveSize) => {
   const { main, cross } = axes[zone.flow]
   /* TODO: The slave is hard-coded to align cross-center with master. */
-  const crossAlign = `center`
-  const mainStart = place(zone.flow, `main`, zone.side, masterBounds, slaveSize)
+  const crossAlign = "center"
+  const mainStart = place(zone.flow, "main", zone.side, masterBounds, slaveSize)
   const mainSize = slaveSize[main.size]
-  const crossStart = place(zone.flow, `cross`, crossAlign, masterBounds, slaveSize)
+  const crossStart = place(zone.flow, "cross", crossAlign, masterBounds, slaveSize)
   const crossSize = slaveSize[cross.size]
 
   return {
@@ -229,7 +249,22 @@ const calcRelPos = (zone, masterBounds, slaveSize) => {
 
 
 
+export default {
+  El,
+  types,
+  validTypeValues,
+  calcRelPos,
+  place,
+  pickZone,
+  axes,
+  centerOfSize,
+  centerOfBounds,
+  centerOfBoundsFromBounds,
+  doesFitWithin,
+  equalCoords: equalRecords,
+}
 export {
+  El,
   types,
   validTypeValues,
   calcRelPos,
@@ -241,5 +276,4 @@ export {
   centerOfBoundsFromBounds,
   doesFitWithin,
   equalRecords as equalCoords,
-  El,
 }
