@@ -1,7 +1,7 @@
 /* eslint no-param-reassign: 0 */
 
-import { window, isServer } from './platform'
-import { noop } from './utils'
+import { window, isServer } from "./platform"
+import { noop } from "./utils"
 
 
 
@@ -21,7 +21,7 @@ const cancelAnimationFrame = isServer
 
 const isIE = isServer ? false : navigator.userAgent.match(/Trident/)
 
-const namespace = `__resizeDetector__`
+const namespace = "__resizeDetector__"
 
 
 
@@ -33,12 +33,12 @@ const uninitialize = (el) => {
 
 
 const createElementHack = () => {
-  const el = document.createElement(`object`)
-  el.className = `resize-sensor`
-  el.setAttribute(`style`, `display: block; position: absolute; top: 0; left: 0; height: 100%; width: 100%; overflow: hidden; pointer-events: none; z-index: -1;`)
-  el.setAttribute(`class`, `resize-sensor`)
-  el.type = `text/html`
-  el.data = `about:blank`
+  const el = document.createElement("object")
+  el.className = "resize-sensor"
+  el.setAttribute("style", "display: block; position: absolute; top: 0; left: 0; height: 100%; width: 100%; overflow: hidden; pointer-events: none; z-index: -1;")
+  el.setAttribute("class", "resize-sensor")
+  el.type = "text/html"
+  el.data = "about:blank"
   return el
 }
 
@@ -61,21 +61,21 @@ const initialize = (el) => {
   if (isIE) {
     /* We do not support ie8 and below (or ie9 in compat mode).
     Therefore there is no presence of `attachEvent` here. */
-    el.addEventListener(`onresize`, onResize)
+    el.addEventListener("onresize", onResize)
     detector.destroy = () => {
-      el.removeEventListener(`onresize`, onResize)
+      el.removeEventListener("onresize", onResize)
     }
   } else {
-    if (getComputedStyle(el).position === `static`) {
+    if (getComputedStyle(el).position === "static") {
       detector.elWasStaticPosition = true
-      el.style.position = `relative`
+      el.style.position = "relative"
     }
     const objEl = createElementHack()
     objEl.onload = function (/* event */) {
-      this.contentDocument.defaultView.addEventListener(`resize`, onResize)
+      this.contentDocument.defaultView.addEventListener("resize", onResize)
     }
     detector.destroy = () => {
-      if (detector.elWasStaticPosition) el.style.position = ``
+      if (detector.elWasStaticPosition) el.style.position = ""
       // Event handlers will be automatically removed.
       // http://stackoverflow.com/questions/12528049/if-a-dom-element-is-removed-are-its-listeners-also-removed-from-memory
       el.removeChild(objEl)
@@ -93,7 +93,7 @@ const on = (el, fn) => {
   special case here so that users do not have to think about two APIs. */
 
   if (el === window) {
-    window.addEventListener(`resize`, fn)
+    window.addEventListener("resize", fn)
     return
   }
 
@@ -107,7 +107,7 @@ const on = (el, fn) => {
 
 const off = (el, fn) => {
   if (el === window) {
-    window.removeEventListener(`resize`, fn)
+    window.removeEventListener("resize", fn)
     return
   }
   const detector = el[namespace]
@@ -119,6 +119,12 @@ const off = (el, fn) => {
 
 
 
+export default {
+  on,
+  off,
+  addEventListener: on,
+  removeEventListener: off,
+}
 export {
   on,
   off,
