@@ -1,7 +1,7 @@
-import { DOM as E, unmountComponentAtNode } from 'react'
-import { render } from 'react-dom'
-import { noop } from './utils'
-import { isClient } from './platform'
+import { DOM as E } from "react"
+import ReactDOM from "react-dom"
+import { noop } from "./utils"
+import { isClient } from "./platform"
 
 
 
@@ -20,7 +20,7 @@ const ReactLayerMixin = () => ({
   componentWillMount () {
     this.targetBounds = null
     /* Create a DOM node for mounting the React Layer. */
-    this.layerContainerNode = createElement(`div`)
+    this.layerContainerNode = createElement("div")
   },
   componentDidMount () {
     /* Mount the mount. */
@@ -39,14 +39,14 @@ const ReactLayerMixin = () => ({
     const layerReactEl = this.renderLayer()
     if (!layerReactEl) {
       this.layerReactComponent = null
-      render(E.noscript(), this.layerContainerNode)
+      ReactDOM.unstable_renderSubtreeIntoContainer(this, E.noscript(), this.layerContainerNode)
     } else {
-      this.layerReactComponent = render(layerReactEl, this.layerContainerNode)
+      this.layerReactComponent = ReactDOM.unstable_renderSubtreeIntoContainer(this, layerReactEl, this.layerContainerNode)
     }
   },
   _layerUnrender () {
     if (this.layerWillUnmount) this.layerWillUnmount(this.layerContainerNode)
-    unmountComponentAtNode(this.layerContainerNode)
+    ReactDOM.unmountComponentAtNode(this.layerContainerNode)
   },
   // renderLayer() {
   //   Must be implemented by consumer.
@@ -55,6 +55,4 @@ const ReactLayerMixin = () => ({
 
 
 
-export {
-  ReactLayerMixin as default,
-}
+export default ReactLayerMixin
