@@ -1,9 +1,6 @@
-import AssignPolyFill from 'object.assign/polyfill'
-import { isClient } from './platform'
+import { isClient } from "./platform"
 
 
-
-const assign = AssignPolyFill()
 
 const arrayify = (x) => (
   Array.isArray(x) ? x : [x]
@@ -28,8 +25,14 @@ const clientOnly = (f) => (
 
 
 
+export default {
+  arrayify,
+  find,
+  equalRecords,
+  noop,
+  clientOnly,
+}
 export {
-  assign,
   arrayify,
   find,
   equalRecords,
